+++ IcVibracoes.Core/Mapper/BeamProfiles/IProfileMapper.cs
@@ -2,7 +2,7 @@
 using IcVibracoes.Core.Models.BeamCharacteristics;
 using System.Threading.Tasks;
 
-namespace IcVibracoes.Core.Mapper.Profiles
+namespace IcVibracoes.Core.Mapper.BeamProfiles
 {
     /// <summary>
     /// It's responsible to build the profile.