+++ IcVibracoes.Core/Mapper/BeamProfiles/Rectangular/IRectangularProfileMapper.cs
@@ -1,6 +1,7 @@
 ﻿using IcVibracoes.Common.Profiles;
+using IcVibracoes.Core.Mapper.BeamProfiles;
 
-namespace IcVibracoes.Core.Mapper.Profiles.Rectangular
+namespace IcVibracoes.Core.Mapper.BeamProfiles.Rectangular
 {
     /// <summary>
     /// It's responsible to build a rectangular profile.