+++ IcVibracoes.Core/NumericalIntegrationMethods/FiniteElement/Newmark/BeamWithDva/BeamWithDvaNewmarkMethod.cs
@@ -3,7 +3,7 @@ using IcVibracoes.Core.Calculator.ArrayOperations;
 using IcVibracoes.Core.DTO.Input;
 using System.Threading.Tasks;
 
-namespace IcVibracoes.Core.NumericalIntegrationMethods.Newmark.BeamWithDva
+namespace IcVibracoes.Core.NumericalIntegrationMethods.FiniteElement.Newmark.BeamWithDva
 {
     /// <summary>
     /// It's responsible to calculate the vibration in a beam with dynamic vibration absorbers.
@@ -23,7 +23,7 @@ namespace IcVibracoes.Core.NumericalIntegrationMethods.Newmark.BeamWithDva
             IAuxiliarOperation auxiliarOperation)
             : base(arrayOperation, auxiliarOperation)
         {
-            this._arrayOperation = arrayOperation;
+            _arrayOperation = arrayOperation;
         }