+++ IcVibracoes.Core/NumericalIntegrationMethods/FiniteElement/Newmark/BeamWithDva/IBeamWithDvaNewmarkMethod.cs
@@ -1,4 +1,4 @@
-﻿namespace IcVibracoes.Core.NumericalIntegrationMethods.Newmark.BeamWithDva
+﻿namespace IcVibracoes.Core.NumericalIntegrationMethods.FiniteElement.Newmark.BeamWithDva
 {
     /// <summary>
     /// It's responsible to calculate the vibration in a beam with dynamic vibration absorbers.