+++ IcVibracoes.Core/NumericalIntegrationMethods/FiniteElement/Newmark/INewmarkMethod.cs
@@ -2,7 +2,7 @@
 using IcVibracoes.DataContracts.FiniteElements;
 using System.Threading.Tasks;
 
-namespace IcVibracoes.Core.NumericalIntegrationMethods.Newmark
+namespace IcVibracoes.Core.NumericalIntegrationMethods.FiniteElement.Newmark
 {
     /// <summary>
     /// It's responsible to execute the Newmark numerical integration method to calculate the vibration.