+++ IcVibracoes.Core/NumericalIntegrationMethods/FiniteElement/NewmarkBeta/NewmarkBetaMethod.cs
@@ -1,0 +1,6 @@
+﻿namespace IcVibracoes.Core.NumericalIntegrationMethods.FiniteElement.NewmarkBeta
+{
+    public class NewmarkBetaMethod : INewmarkBetaMethod
+    {
+    }
+}