+++ IcVibracoes.Core/NumericalIntegrationMethods/NewmarkBeta/INewmarkBetaMethod.cs
@@ -1,6 +1,0 @@
-﻿namespace IcVibracoes.Core.NumericalIntegrationMethods.NewmarkBeta
-{
-    public interface INewmarkBetaMethod
-    {
-    }
-}