+++ IcVibracoes.Core/Operations/FiniteElements/CalculateVibration/BeamWithDva/Rectangular/CalculateRectangularBeamWithDvaVibration.cs
@@ -4,8 +4,8 @@ using IcVibracoes.Core.Calculator.ArrayOperations;
 using IcVibracoes.Core.Calculator.MainMatrixes.Beam.Rectangular;
 using IcVibracoes.Core.Calculator.MainMatrixes.BeamWithDva.Rectangular;
 using IcVibracoes.Core.Mapper;
-using IcVibracoes.Core.Mapper.Profiles.Rectangular;
-using IcVibracoes.Core.NumericalIntegrationMethods.Newmark.BeamWithDva;
+using IcVibracoes.Core.Mapper.BeamProfiles.Rectangular;
+using IcVibracoes.Core.NumericalIntegrationMethods.FiniteElement.Newmark.BeamWithDva;
 using IcVibracoes.Core.Validators.Profiles.Rectangular;
 
 namespace IcVibracoes.Core.Operations.FiniteElements.CalculateVibration.BeamWithDva.Rectangular