+++ IcVibracoes.Core/Operations/FiniteElements/CalculateVibration/BeamWithPiezoelectric/Circular/CalculateCircularBeamWithPiezoelectricVibration.cs
@@ -4,9 +4,9 @@ using IcVibracoes.Core.Calculator.ArrayOperations;
 using IcVibracoes.Core.Calculator.MainMatrixes.BeamWithPiezoelectric.Circular;
 using IcVibracoes.Core.Mapper;
 using IcVibracoes.Core.Mapper.PiezoelectricProfiles.Circular;
-using IcVibracoes.Core.Mapper.Profiles.Circular;
-using IcVibracoes.Core.NumericalIntegrationMethods.Newmark;
+using IcVibracoes.Core.Mapper.BeamProfiles.Circular;
 using IcVibracoes.Core.Validators.Profiles.Circular;
+using IcVibracoes.Core.NumericalIntegrationMethods.FiniteElement.Newmark;
 
 namespace IcVibracoes.Core.Operations.FiniteElements.CalculateVibration.BeamWithPiezoelectric.Circular
 {