+++ IcVibracoes/Startup.cs
@@ -1,5 +1,7 @@
 using IcVibracoes.Calculator.GeometricProperties;
 using IcVibracoes.Core.AuxiliarOperations;
+using IcVibracoes.Core.AuxiliarOperations.Eigenvalue;
+using IcVibracoes.Core.AuxiliarOperations.NaturalFrequency;
 using IcVibracoes.Core.Calculator.ArrayOperations;
 using IcVibracoes.Core.Calculator.MainMatrixes.Beam.Circular;
 using IcVibracoes.Core.Calculator.MainMatrixes.Beam.Rectangular;
@@ -8,12 +10,13 @@ using IcVibracoes.Core.Calculator.MainMatrixes.BeamWithDva.Rectangular;
 using IcVibracoes.Core.Calculator.MainMatrixes.BeamWithPiezoelectric.Circular;
 using IcVibracoes.Core.Calculator.MainMatrixes.BeamWithPiezoelectric.Rectangular;
 using IcVibracoes.Core.Mapper;
+using IcVibracoes.Core.Mapper.BeamProfiles.Circular;
+using IcVibracoes.Core.Mapper.BeamProfiles.Rectangular;
 using IcVibracoes.Core.Mapper.PiezoelectricProfiles.Circular;
 using IcVibracoes.Core.Mapper.PiezoelectricProfiles.Rectangular;
-using IcVibracoes.Core.Mapper.Profiles.Circular;
-using IcVibracoes.Core.Mapper.Profiles.Rectangular;
-using IcVibracoes.Core.NumericalIntegrationMethods.Newmark;
-using IcVibracoes.Core.NumericalIntegrationMethods.Newmark.BeamWithDva;
+using IcVibracoes.Core.NumericalIntegrationMethods.FiniteElement.Newmark;
+using IcVibracoes.Core.NumericalIntegrationMethods.FiniteElement.Newmark.BeamWithDva;
+using IcVibracoes.Core.NumericalIntegrationMethods.FiniteElement.NewmarkBeta;
 using IcVibracoes.Core.NumericalIntegrationMethods.RigidBody.RungeKuttaForthOrder.OneDegreeFreedom;
 using IcVibracoes.Core.NumericalIntegrationMethods.RigidBody.RungeKuttaForthOrder.TwoDegreeFreedom;
 using IcVibracoes.Core.Operations.FiniteElements.CalculateVibration.Beam.Circular;
@@ -26,6 +29,7 @@ using IcVibracoes.Core.Operations.RigidBody.CalculateVibration.OneDegreeFreedom;
 using IcVibracoes.Core.Operations.RigidBody.CalculateVibration.TwoDegreesFreedom;
 using IcVibracoes.Core.Validators.Profiles.Circular;
 using IcVibracoes.Core.Validators.Profiles.Rectangular;
+using IcVibracoes.Core.Validators.TimeStep;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
@@ -47,55 +51,75 @@ namespace IcVibracoes
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
-            // Calculator - ArrayOperation
+            // Auxiliar Operations
+            services.AddScoped<ICalculateEigenvalue, CalculateEigenvalue>();
+            services.AddScoped<INaturalFrequency, NaturalFrequency>();
+            services.AddScoped<IAuxiliarOperation, AuxiliarOperation>();
+
+            // Calculator - Array Operation
             services.AddScoped<IArrayOperation, ArrayOperation>();
-            // Calculator - CalculateGeometricProperty
+
+            // Calculator - Geometric Property
             services.AddScoped<ICalculateGeometricProperty, CalculateGeometricProperty>();
-            // Calculator - MainMatrix
-            services.AddScoped<IRectangularBeamMainMatrix, RectangularBeamMainMatrix>();
+
+            // Calculator - Main Matrixes - Beam
             services.AddScoped<ICircularBeamMainMatrix, CircularBeamMainMatrix>();
+            services.AddScoped<IRectangularBeamMainMatrix, RectangularBeamMainMatrix>();
+
+            // Calculator - Main Matrixes - Beam with Dynamic Vibration Absorber
             services.AddScoped<ICircularBeamWithDvaMainMatrix, CircularBeamWithDvaMainMatrix>();
             services.AddScoped<IRectangularBeamWithDvaMainMatrix, RectangularBeamWithDvaMainMatrix>();
-            services.AddScoped<IRectangularBeamWithPiezoelectricMainMatrix, RectangularBeamWithPiezoelectricMainMatrix>();
+
+            // Calculator - Main Matrixes - Beam with Piezoelectric
             services.AddScoped<ICircularBeamWithPiezoelectricMainMatrix, CircularBeamWithPiezoelectricMainMatrix>();
+            services.AddScoped<IRectangularBeamWithPiezoelectricMainMatrix, RectangularBeamWithPiezoelectricMainMatrix>();
 
             // Mapper
             services.AddScoped<IMappingResolver, MappingResolver>();
+            
+            // Mapper - Beam Profiles
             services.AddScoped<ICircularProfileMapper, CircularProfileMapper>();
             services.AddScoped<IRectangularProfileMapper, RectangularProfileMapper>();
+
+            // Mapper - Piezoelectric Profiles
             services.AddScoped<IPiezoelectricCircularProfileMapper, PiezoelectricCircularProfileMapper>();
             services.AddScoped<IPiezoelectricRectangularProfileMapper, PiezoelectricRectangularProfileMapper>();
-
-            // Newmark Numerical Integration
+            
+            // Numerical Integration Methods - Finite Element - Newmark
             services.AddScoped<INewmarkMethod, NewmarkMethod>();
             services.AddScoped<IBeamWithDvaNewmarkMethod, BeamWithDvaNewmarkMethod>();
 
-            // Runge Kutta Forth Order Numerical Integration
+            // Numerical Integration Methods - Finite Element - Newmark Beta
+            services.AddScoped<INewmarkBetaMethod, NewmarkBetaMethod>();
+
+            // Numerical Integration Methods - Rigid Body - Runge Kutta Forth Order
             services.AddScoped<IRungeKuttaForthOrderMethod_1DF, RungeKuttaForthOrderMethod_1DF>();
             services.AddScoped<IRungeKuttaForthOrderMethod_2DF, RungeKuttaForthOrderMethod_2DF>();
-
-            // Auxiliar Operations
-            services.AddScoped<IAuxiliarOperation, AuxiliarOperation>();
             
             // Rigid Body Operations
             services.AddScoped<ICalculateVibrationToOneDegreeFreedom, CalculateVibrationToOneDegreeFreedom>();
             services.AddScoped<ICalculateVibrationToTwoDegreesFreedom, CalculateVibrationToTwoDegreesFreedom>();
 
-            // Beam Operations
+            // Finite Element Operations - Beam
             services.AddScoped<ICalculateCircularBeamVibration, CalculateCircularBeamVibration>();
             services.AddScoped<ICalculateRectangularBeamVibration, CalculateRectangularBeamVibration>();
 
-            // BeamWithDva Operations
+            // Finite Element Operations - Beam with Dynamic Vibration Absorber
             services.AddScoped<ICalculateCircularBeamWithDvaVibration, CalculateCircularBeamWithDvaVibration>();
             services.AddScoped<ICalculateRectangularBeamWithDvaVibration, CalculateRectangularBeamWithDvaVibration>();
 
-            // Piezoelectric Operations
+            // Finite Element Operations - Beam with Piezoelectric
             services.AddScoped<ICalculateCircularBeamWithPiezoelectricVibration, CalculateCircularBeamWithPiezoelectricVibration>();
             services.AddScoped<ICalculateRectangularBeamWithPiezoelectricVibration, CalculateRectangularBeamWithPiezoelectricVibration>();
 
-            // Validators
+            // Validators -  Beam Request Data
+
+            // Validators - Profiles
             services.AddScoped<IRectangularProfileValidator, RectangularProfileValidator>();
             services.AddScoped<ICircularProfileValidator, CircularProfileValidator>();
+
+            // Validators - Time Step
+            services.AddScoped<ITimeStepValidator, TimeStepValidator>();
 
             services.AddControllers();